TicTacToe v1.0:
Features:
Click to play the current turn
When all boards are filled up, a new board is created.
When a tic-tac-toe is completed, a new board is created.
When a new board cannot be created, it is allocated to the top right.
When a player gets 3 congruent tic-tac-toes (corners count too), they
win
Press Enter to restart.